requirements.txt
pyTelegramBotAPI==4.14.1
requests
qrcode
pillow
python-dotenv
schedule

